requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
beautifulsoup4
certifi
charset-normalizer
distro
frozenlist
h11
httpcore
httpx
idna
jiter
multidict
numpy
opencv-python-headless
packaging
pillow
pydantic
pydantic_core
pytesseract
python-dotenv
requests
sniffio
soupsieve
tqdm
typing_extensions
urllib3
yarl
youtube-transcript-api
notdiamond[create]
litellm
tk
streamlit